refactor : 콘솔 컨트롤러 클래스명 변경 및 view package -> console package로 변경
