+++ src/main/java/kr/co/springbootweeklymission/console/Command.java
@@ -1,4 +1,4 @@
-package kr.co.springbootweeklymission.view;
+package kr.co.springbootweeklymission.console;
 
 import kr.co.springbootweeklymission.common.error.exception.NotFoundException;
 import kr.co.springbootweeklymission.common.response.ResponseStatus;
